fixed update and delete news
